requirements.txt
blinker==1.6.2
click==8.1.4
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mrz==0.6.2
numpy==1.25.0
opencv-contrib-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
pytesseract==0.3.10
SQLAlchemy==2.0.18
typing_extensions==4.7.1
Werkzeug==2.3.6
